user will be able to bid on a running auction

+++ auction-service/src/main/java/com/biddingserver/auctionservice/model/BidRequestDTO.java
@@ -7,5 +7,5 @@ import lombok.*;
 @AllArgsConstructor
 public class BidRequestDTO
 {
-    Integer bidAmount;
+    Long bidAmount;
 }
